Initialize wheel slip parameters directly from SDF
Do not rely on the parameter_events topic for initialization.
Otherwise, it's easy to race with other nodes that want to query or update the wheel slip parameters.

This commit changes the behavior of the parameters affecting all wheels.
These parameters are no longer initialized if a user does not provide a value, i.e. they are
'unset' until a value is set. Previously, they defaulted to the value set on an individual wheel.

For purposes of initializing values, parameters setting individual wheels now have higher precendence
than the parameters for all wheels.

I also bumped the number of retries (timeout) in the tests, since I noticed a larger delay with certain
RMWs (ie. rmw_connextdds).

Signed-off-by: Jacob Perron <jacob@openrobotics.org>